# api/openapi.yaml
openapi: 3.0.3
info:
  title: Writing Assistant Engine API
  description: >
    The academic writing-assistant engine's REST/SSE surface (ADR-0017). Typed
    non-streaming routes are codegen'd with ogen; the `/turn` SSE stream is
    hand-framed in the API server (ADR-0031) and is marked x-ogen-raw-response so
    ogen hands the handler the raw http.ResponseWriter.
  version: 0.1.0
servers:
  - url: http://127.0.0.1:9100
paths:
...
  /models/{name}/start:
    post:
      operationId: startModel
      parameters:
        - $ref: '#/components/parameters/Name'
      responses:
        '200':
          description: started (up)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveStateResponse'
  /models/{name}/stop:
    post:
      operationId: stopModel
      parameters:
        - $ref: '#/components/parameters/Name'
      responses:
        '200':
          description: stopped (idempotent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveStateResponse'
  /models/{name}/provision:
    post:
      operationId: provisionModel
      parameters:
        - $ref: '#/components/parameters/Name'
      responses:
        '202':
          description: accepted (async download)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvisionResponse'
  /models/{name}/status:
    get:
      operationId: getModelStatus
      parameters:
        - $ref: '#/components/parameters/Name'
      responses:
        '200':
          description: live state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveStateResponse'
...
  /directories:
    get:
      operationId: listDirectory
      summary: List one directory's direct, non-recursive entries (ADR-0035)
      parameters:
        - name: path
          in: query
          required: true
          description: absolute path of the directory to list
          schema:
            type: string
      responses:
        '200':
          description: the directory's entries (shallow, name-sorted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryListing'
  /documents:
    post:
      operationId: openDocument
      summary: Open/create a document by path; returns its surrogate id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenDocumentRequest'
      responses:
        '200':
          description: document id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
  /documents/{id}/blocks:
    get:
      operationId: getBlocks
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: block tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Block'
  /documents/{id}/edits:
    post:
      operationId: applyEdit
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockEdit'
      responses:
        '200':
          description: revision (candidate staged)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revision'
  /documents/{id}/commits:
    post:
      operationId: commitDocument
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revision'
  /documents/{id}/tree:
    put:
      operationId: saveDocument
      summary: Replace the document's block tree (manual autosave, ADR-0038)
      description: >
        The manual-edit wire path (ADR-0038): the client's whole block-tree
        snapshot. Array order = position; a block with no `id` is new (the engine
        mints a UUID, ADR-0020 §3); an `id` present in the current tree but absent
        from the request is deleted; a changed `kind`/`parentId` is retyped/moved.
        The engine reconciles, normalizes on write and formats on commit, and
        commits `autosave @ <ts>` iff anything changed (a no-op returns the current
        HEAD). A manual save of a block drops its open candidates.
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveTreeRequest'
      responses:
        '200':
          description: resulting revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revision'
  /documents/{id}/history:
    get:
      operationId: getHistory
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: revisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Revision'
  /documents/{id}/diff:
    get:
      operationId: getDiff
      parameters:
        - $ref: '#/components/parameters/DocumentID'
        - name: base
          in: query
          required: true
          schema:
            type: string
        - name: rev
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: word edits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WordEdit'
  /documents/{id}/blocks/{bid}/candidates:
    get:
      operationId: getCandidates
      parameters:
        - $ref: '#/components/parameters/DocumentID'
        - $ref: '#/components/parameters/BlockID'
      responses:
        '200':
          description: open candidates for one block
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Candidate'
  /turn:
    post:
      operationId: startTurn
      summary: Run one agent turn (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: >
            SSE event stream. Each message is `event: <type>` followed by
            `data: <payload>` where <type> is one of the Event.type enum values
            and <payload> is the matching component schema (TokenEvent,
            MeterEvent, CandidateEvent, DiffEvent, RagEvent, DoneEvent,
            ErrorEvent, BackpressureEvent). One turn per connection: the server
            demultiplexes a turn's events to exactly one client stream, so
            payloads do not repeat the turnId. (ADR-0017 §6, amended.)
          content:
            text/event-stream:
              x-ogen-raw-response: true
              schema:
                $ref: '#/components/schemas/Event'
  /sessions:
    get:
      operationId: listSessions
      parameters:
        - name: documentId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: sessions for a document
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
    post:
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: created-or-resumed session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
  /sessions/{id}/messages:
    get:
      operationId: getSessionMessages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: message history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
components:
  parameters:
    Name:
      name: name
      in: path
      required: true
      schema:
        type: string
    DocumentID:
      name: id
      in: path
      required: true
      schema:
        type: string
    BlockID:
      name: bid
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    LiveStateResponse:
      type: object
      required: [name, state]
      properties:
        name:
          type: string
        state:
          type: string
          enum: [up, down, starting, stopping, provisioning, unknown]
    ProvisionResponse:
      type: object
      required: [provisionID]
      properties:
        provisionID:
          type: string
...
    OpenDocumentRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
    Document:
      type: object
      required: [id, path, rootBlockId]
      properties:
        id:
          type: string
        path:
          type: string
        rootBlockId:
          type: string
        updatedAt:
          type: integer
          format: int64
    Entry:
      type: object
      description: One directory entry (ADR-0035). Hidden entries are returned; display filtering is client-side.
      required: [name, path, isDir]
      properties:
        name:
          type: string
        path:
          type: string
        isDir:
          type: boolean
    DirectoryListing:
      type: object
      description: A shallow, non-recursive directory listing (ADR-0035 §2).
      required: [path, entries]
      properties:
        path:
          type: string
        entries:
          type: array
          items:
            $ref: '#/components/schemas/Entry'
    Block:
      type: object
      required: [id, kind, position, text]
      properties:
        id:
          type: string
        parentId:
          type: string
        kind:
          type: string
          enum: [paragraph, heading, list_item, code_fence, blockquote, table]
        position:
          type: integer
        text:
          type: string
        hash:
          type: string
    Guard:
      type: object
      required: [blockId, hash]
      properties:
        blockId:
          type: string
        hash:
          type: string
    BlockEdit:
      type: object
      required: [blockId, text]
      properties:
        blockId:
          type: string
        text:
          type: string
        guards:
          type: array
          items:
            $ref: '#/components/schemas/Guard'
    SaveTreeRequest:
      type: object
      description: The manual-edit whole-tree snapshot (ADR-0038). Array order = position.
      required: [blocks]
      properties:
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/BlockWrite'
    BlockWrite:
      type: object
      description: >
        One block in a manual tree save. `id` absent = new block (engine mints a
        UUID, ADR-0020 §3); `id` present = existing block matched by stable UUID.
        No `hash`/`guards` — those are AI-edit concerns (ADR-0029 §4).
      required: [kind, text]
      properties:
        id:
          type: string
        parentId:
          type: string
        kind:
          type: string
          enum: [paragraph, heading, list_item, code_fence, blockquote, table]
        text:
          type: string
    Revision:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        timestamp:
          type: integer
          format: int64
    WordEdit:
      type: object
      required: [blockId]
      properties:
        blockId:
          type: string
        insertions:
          type: array
          items:
            type: string
        deletions:
          type: array
          items:
            type: string
    Candidate:
      type: object
      properties:
        blockId:
          type: string
        text:
          type: string
        baseId:
          type: string
    Selection:
      type: object
      properties:
        blockId:
          type: string
    TurnOptions:
      type: object
      properties:
        temperature:
          type: number
        model:
          type: string
    Mention:
      type: object
      description: A turn-scoped file attachment (ADR-0036); absolute path resolved by the client.
      required: [path]
      properties:
        path:
          type: string
    Task:
      type: object
      required: [sessionId, modeName, documentId, userInput]
      properties:
        sessionId:
          type: string
        modeName:
          type: string
        documentId:
          type: string
        userInput:
          type: string
        selection:
          $ref: '#/components/schemas/Selection'
        mentions:
          type: array
          items:
            $ref: '#/components/schemas/Mention'
        options:
          $ref: '#/components/schemas/TurnOptions'
    Message:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, tool]
        content:
          type: string
        timestamp:
          type: integer
          format: int64
    Session:
      type: object
      required: [id, documentId]
      properties:
        id:
          type: string
        documentId:
          type: string
        anchorBlockId:
          type: string
        modeType:
          type: string
        title:
          type: string
        tokenBudget:
          type: integer
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
    CreateSessionRequest:
      type: object
      required: [documentId]
      properties:
        documentId:
          type: string
        anchorBlockId:
          type: string
        modeType:
          type: string

    # SSE event vocabulary (ADR-0017 §6, amended Track 1 TUI session). The
    # hand-framed server writes `event: <type>` + `data: <payload>`; the client
    # codegens (Hey API + Zod, openapi-to-rust) consume the payload schemas
    # below as typed SSE decoders keyed by the SSE event name. turnId
    # correlation is per-connection (one turn per stream) and is not repeated
    # in payloads.
    Event:
      type: object
      description: >
        Framing marker for the SSE stream: the `event:` line carries `type`;
        the `data:` line carries the payload schema matching that type. The
        payload JSON itself has no `type` field.
      required: [type]
      properties:
        type:
          type: string
          enum: [token, meter, candidate, diff, rag, done, error, backpressure]
...